+++ CHANGELOG.md
@@ -1,11 +1,13 @@
 # CHANGELOG
 
-## v25.47
+
 
 ## v25.xx
 
 ### Features
-- New output notifications. (#117 v25.50-alpha) 
+- New output push notifications. (#117 v25.50-alpha)
+- Push notification supports tenant configuration. (#121 v26.02-alpha)
+## v25.47
 
 ### Features
 - Jinja template support for error notifications. (#112 v25.42-alpha2)